Move _keyPairToDidDocument to helpers.
