+++ lib/DidWebDriver.js
@@ -5,31 +5,16 @@ import * as didIo from '@digitalbazaar/did-io';
 import {
   contextsBySuite,
   defaultFetchOptions,
-  DID_CONTEXT_URL,
-  ED25519_KEY_2018_CONTEXT_URL
 } from './constants.js';
 import {
   didToUrl,
+  keyPairToDidDocument,
   urlToDid
 } from './helpers.js';
 import {
   Ed25519VerificationKey2020
 } from '@digitalbazaar/ed25519-verification-key-2020';
 import {httpClient} from '@digitalbazaar/http-client';
-import {
-  X25519KeyAgreementKey2019
-} from '@digitalbazaar/x25519-key-agreement-key-2019';
-import {
-  X25519KeyAgreementKey2020
-} from '@digitalbazaar/x25519-key-agreement-key-2020';
-
-/**
- * A Linked Data Key Pair.
- *
- * @typedef {object} LDKeyPair
- * @see https://github.com/digitalbazaar/crypto-ld/blob/main/lib/LDKeyPair.js
- *
-*/
 
 export class DidWebDriver {
   /**
@@ -74,9 +59,10 @@ export class DidWebDriver {
     // keyPairs is a map of keyId to key pair instance, that includes
     // the verificationKeyPair above, but also the keyAgreementKey pair that
     // is derived from the verification key pair.
-    const {didDocument, keyPairs} = await this._keyPairToDidDocument({
+    const {didDocument, keyPairs} = await keyPairToDidDocument({
       did: urlToDid(url),
-      keyPair: verificationKeyPair
+      keyPair: verificationKeyPair,
+      verificationSuite: this.verificationSuite
     });
 
     // Convenience function that returns the public/private key pair instance
@@ -157,73 +143,6 @@ export class DidWebDriver {
     const didDocument = await httpClient.get(didToUrl(did), requestOptions);
     // Resolve the full DID Document
     return didDocument;
-  }
-  /**
-   * Converts an Ed25519KeyPair object to a `did:web` method DID Document.
-   *
-   * @param {object} options - Options object.
-   * @param {LDKeyPair|object} options.keyPair - Key used to generate the DID
-   *   document (either an LDKeyPair instance containing public key material,
-   *   or a "key description" plain object (such as that generated from a KMS)).
-   * @param {string} options.did - The did for the did document.
-   *
-   * @returns {Promise<{didDocument: object, keyPairs: Map}>}
-   *   Resolves with the generated DID Document, along with the corresponding
-   *   key pairs used to generate it (for storage in a KMS).
-   */
-  async _keyPairToDidDocument({keyPair, did} = {}) {
-    const verificationKeyPair = await this.verificationSuite.from({...keyPair});
-    verificationKeyPair.controller = did;
-
-    const contexts = [DID_CONTEXT_URL];
-
-    // The KAK pair will use the source key's controller, but will generate
-    // its own .id
-    let keyAgreementKeyPair;
-    if(verificationKeyPair.type === 'Ed25519VerificationKey2020') {
-      keyAgreementKeyPair = X25519KeyAgreementKey2020
-        .fromEd25519VerificationKey2020({keyPair: verificationKeyPair});
-      contexts.push(Ed25519VerificationKey2020.SUITE_CONTEXT,
-        X25519KeyAgreementKey2020.SUITE_CONTEXT);
-    } else if(verificationKeyPair.type === 'Ed25519VerificationKey2018') {
-      keyAgreementKeyPair = X25519KeyAgreementKey2019
-        .fromEd25519VerificationKey2018({keyPair: verificationKeyPair});
-      contexts.push(ED25519_KEY_2018_CONTEXT_URL,
-        X25519KeyAgreementKey2019.SUITE_CONTEXT);
-    } else {
-      throw new Error(
-        'Cannot derive key agreement key from verification key type "' +
-          verificationKeyPair.type + '".'
-      );
-    }
-
-    // Now set the source key's id
-    verificationKeyPair.id = `${did}#${verificationKeyPair.fingerprint()}`;
-
-    // get the public components of each keypair
-    const publicEdKey = verificationKeyPair.export({publicKey: true});
-    const publicDhKey = keyAgreementKeyPair.export({publicKey: true});
-
-    // Compose the DID Document
-    const didDocument = {
-      // Note that did:web does not have its own method-specific context,
-      // and only uses the general DID Core context, and key-specific contexts.
-      '@context': contexts,
-      id: did,
-      verificationMethod: [publicEdKey],
-      authentication: [publicEdKey.id],
-      assertionMethod: [publicEdKey.id],
-      capabilityDelegation: [publicEdKey.id],
-      capabilityInvocation: [publicEdKey.id],
-      keyAgreement: [publicDhKey]
-    };
-
-    // create the key pairs map
-    const keyPairs = new Map();
-    keyPairs.set(verificationKeyPair.id, verificationKeyPair);
-    keyPairs.set(keyAgreementKeyPair.id, keyAgreementKeyPair);
-
-    return {didDocument, keyPairs};
   }
 }
